backend/requirements.txt
# requirements.txt

# Django framework for the web application
Django==5.0.7

# Django REST Framework for building the API
djangorestframework==3.16.0

# Library to handle environment variables for secure settings
# Using the python-decouple package is a simple, effective solution.
python-decouple==3.8

# CORS headers for handling cross-origin requests from the frontend
django-cors-headers==4.9.0

# Database connector for PostgreSQL
psycopg2-binary==2.9.9

# Asynchronous task queue for handling background processes
# We'll include the Redis transport dependency here.
celery[redis]==5.4.0

# Redis client for Python, used by Celery and as a cache backend
redis==5.0.1

# Gunicorn is a popular WSGI HTTP Server for UNIX, used for production deployments
gunicorn==22.0.0